fix: remove password pre-fill in edit form, fix link text and remove redundant cast
Agent-Logs-Url: https://github.com/MeG123321/v1/sessions/36cb9361-ac20-4123-ac3f-7b1e04a43f0e

+++ Magazyn/Controllers/UzytkownicyController.cs
@@ -449,7 +449,7 @@ ORDER BY DataZapomnienia DESC;
         using var con = Db.OpenConnection(DbPath);
 
         static int SecureNext(int maxExclusive) =>
-            (int)(RandomNumberGenerator.GetInt32(maxExclusive));
+            RandomNumberGenerator.GetInt32(maxExclusive);
 
         static string RandDigits(int len)
         {
